test/example/requirements.txt
requests>=2.10.0
sseclient-py>=1.5
